+++ nuxt/components/modal/ModalImageSelection.vue
@@ -34,10 +34,10 @@ getApiData([profilePictureSetMutation])
 // data
 const isTesting = config.public.isTesting
 const routeParamUsername = route.params.username as string
-const selectedProfilePictureStorageKey = ref<string>()
+const selectedProfilePictureStorageKey = ref<string | null>()
 
 // methods
-function selectProfilePictureStorageKey(storageKey?: string) {
+function selectProfilePictureStorageKey(storageKey?: string | null) {
   selectedProfilePictureStorageKey.value = storageKey
 }
 async function setProfilePicture() {